Tweak the `RouterContext` to allow keeping the query string parameters
